refactor: pair down API calls and Dataframes
In the refactor, a lot of unnecessary code was deleted.
Before there were 6 DataFrames that were being created. Now there is only one.
The DataFrame has all games that have been played by the teams (to date).
The games are labeled (boolean values) as to whether they are:

- Regular Season
- Post Trade Deadline
- Post Season
- Head-to-Head Matchups

Introduced the following method:
- get_matchup_df